Reformat and add column to match the new workday format
It goes without saying that the new format is abysmal.
The extra lines at the beginning of each CSV file
just makes it difficult to create these files
(as can be seen from the creation of the excel file).
There are also several columns where the value is the same
for each students, so these could just have been omitted
and autofilled on workday's end.